refactor(desktop): move table into @conar/table package
