+++ packages/table/package.json
@@ -1,0 +1,22 @@
+{
+  "name": "@conar/table",
+  "type": "module",
+  "private": true,
+  "exports": {
+    ".": "./src/index.ts",
+    "./constants": "./src/constants.ts"
+  },
+  "scripts": {
+    "check-types": "tsc"
+  },
+  "dependencies": {
+    "@conar/configs": "workspace:*",
+    "@conar/shared": "workspace:*",
+    "@conar/ui": "workspace:*",
+    "@fluentui/react-context-selector": "catalog:",
+    "@remixicon/react": "catalog:",
+    "@tanstack/react-virtual": "catalog:",
+    "@types/react": "catalog:",
+    "react": "catalog:"
+  }
+}
