+++ packages/table/src/constants.ts
@@ -1,0 +1,2 @@
+export const DEFAULT_ROW_HEIGHT = 32
+export const DEFAULT_COLUMN_WIDTH = 240
