+++ packages/table/src/table-context.ts
@@ -2,7 +2,7 @@ import type { ScrollDirection } from '@conar/ui/hookas/use-scroll-direction'
 import type { ContextSelector } from '@fluentui/react-context-selector'
 import type { VirtualItem } from '@tanstack/react-virtual'
 import type { RefObject } from 'react'
-import type { ColumnRenderer } from '.'
+import type { ColumnRenderer } from './'
 import { createContext, useContextSelector } from '@fluentui/react-context-selector'
 
 export interface TableContextType {
