+++ packages/table/src/use-shift-selection-click.ts
